pages for kittens and about created as seperate components
